refactor: move components and styles to src
